add driverjs tour refactor.
